==> 02 - Annual and Perennial Separation/04 - Merging Results.ipynb <==
# SCRIPT 04 - Merging Results

The previous script made predictions according to each tile separately. The script did not immediately merge all results in order to provide better flexibility for the user to decide the order and amount of tiles to be processed in each run. Therefore, this script helps mosaicing all tiles to generate a single .tif file with all results.

For the correct use of this script, GDAL must be usable from terminal/command prompt. Furthermore, the following libraries are necessary.

+ `glob`
    + Library used to gather file paths according to name and path patterns.
+ `os`
    + Library used to handle paths concatenation.

import glob
import os

The following parameter must be adapted:

+ `path_to_predictions_folder` (`string`)
    + Path to the folder where the predictions for tiles are stored.

path_to_predictions_folder = '/path/to/predictions/folder'

Gather all .tif files inside predictions folder.

files = glob.glob(os.path.join(path_to_predictions_folder, '*.tif'))
files.sort()
len(files)

Creates a .vrt file with all tiles and then use gdal_translate to generate the final file.

os.chdir(path_to_predictions_folder)

with open('./vrtInputFiles.txt', 'w') as f:
    f.write('\n'.join(files))

command = 'gdalbuildvrt -vrtnodata 0 -input_file_list vrtInputFiles.txt aux.vrt'
subprocess.call(command, shell=True)

command = f'gdal_translate aux.vrt merged_predictions.tif -co COMPRESS=PACKBITS'
!{command}
